requirements.txt
feedparser
typer
django
tqdm
toml
openai
python-dateutil
ollama
pydantic